fix: Fixed data type mismatch (int → bigint) in entity and controller.

+++ src/main/java/tfg/warhammerScanner/controller/InfoLinkController.java
@@ -59,7 +59,7 @@ public class InfoLinkController {
     }
 
     @GetMapping("/cost/{idSelection}")
-    public ResponseEntity<Integer> getCostBySelection(@PathVariable Long idSelection) {
+    public ResponseEntity<Long> getCostBySelection(@PathVariable Long idSelection) {
         List<Cost> costs = costRepository.findByIdSelectionAndName(idSelection, "pts");
         if (costs.isEmpty()) {
             return ResponseEntity.notFound().build();
