+++ src/main/java/tfg/warhammerScanner/entity/Cost.java
@@ -18,7 +18,7 @@ public class Cost {
     private String name;
 
     @Column(name="value")
-    private Integer value;
+    private Long value;
 
 
     public Long getIdCost() {
@@ -45,11 +45,11 @@ public class Cost {
         this.name = name;
     }
 
-    public Integer getValue() {
+    public Long getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Long value) {
         this.value = value;
     }
 }
